frontend: extract client helpers from mux.go into client.go
Move standalone types, constants, and functions from mux.go into a
dedicated client.go file. This is a pure code movement with no
behavioral changes, separating client/request helpers from the
BuildMux routing logic to prepare for replacing BuildMux with a
flat Router.
