+++ frontend/mux.go
@@ -10,8 +10,6 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/containerd/platforms"
-	"github.com/goccy/go-yaml"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/moby/buildkit/frontend/dockerui"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
@@ -22,15 +20,6 @@ import (
 	"github.com/project-dalec/dalec"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/exp/maps"
-)
-
-const (
-	keyResolveSpec = "frontend.dalec.resolve"
-
-	// KeyDefaultPlatform is the subreuqest id for returning the default platform
-	// for the builder.
-	KeyDefaultPlatform = "frontend.dalec.defaultPlatform"
-	KeyJSONSchema      = "frontend.dalec.schema"
 )
 
 // BuildMux implements a buildkit BuildFunc via its Handle method. With a
@@ -94,8 +83,6 @@ func (m *BuildMux) Add(targetPath string, bf gwclient.BuildFunc, info *bktargets
 	bklog.G(context.TODO()).WithField("target", targetPath).Info("Added handler to router")
 }
 
-const keyTarget = "target"
-
 // describe returns the subrequests that are supported
 func (m *BuildMux) describe() (*gwclient.Result, error) {
 	subs := []subrequests.Request{bktargets.SubrequestsTargetsDefinition, subrequests.SubrequestsDescribeDefinition, {
@@ -149,70 +136,6 @@ func (m *BuildMux) handleSubrequest(ctx context.Context, client gwclient.Client,
 	}
 }
 
-func handleResolveSpec(ctx context.Context, client gwclient.Client) (*gwclient.Result, error) {
-	dc, err := dockerui.NewClient(client)
-	if err != nil {
-		return nil, err
-	}
-
-	targets := dc.TargetPlatforms
-	if len(targets) == 0 {
-		targets = append(targets, platforms.DefaultSpec())
-	}
-
-	out := make([]*dalec.Spec, 0, len(targets))
-	for _, p := range targets {
-		spec, err := LoadSpec(ctx, dc, &p)
-		if err != nil {
-			return nil, err
-		}
-		out = append(out, spec)
-	}
-
-	dtYaml, err := yaml.Marshal(out)
-	if err != nil {
-		return nil, err
-	}
-
-	dtJSON, err := json.Marshal(out)
-	if err != nil {
-		return nil, err
-	}
-
-	res := gwclient.NewResult()
-	res.AddMeta("result.json", dtJSON)
-	// result.txt here so that `docker buildx build --print` will output it directly.
-	// Otherwise it prints a go object.
-	res.AddMeta("result.txt", dtYaml)
-
-	return res, nil
-}
-
-// handleDefaultPlatform returns the default platform
-func handleDefaultPlatform() (*gwclient.Result, error) {
-	res := gwclient.NewResult()
-
-	p := platforms.DefaultSpec()
-	dt, err := json.Marshal(p)
-	if err != nil {
-		return nil, err
-	}
-
-	res.AddMeta("result.json", dt)
-	res.AddMeta("result.txt", []byte(platforms.Format(p)))
-
-	return res, nil
-}
-
-func handleJSONSchema() (*gwclient.Result, error) {
-	res := gwclient.NewResult()
-	jsonSchema := dalec.GetJSONSchema()
-	res.AddMeta("result.json", jsonSchema)
-	res.AddMeta("result.txt", jsonSchema)
-
-	return res, nil
-}
-
 func (m *BuildMux) loadSpec(ctx context.Context, client gwclient.Client) (*dalec.Spec, error) {
 	if m.spec != nil {
 		return m.spec, nil
@@ -234,24 +157,6 @@ func (m *BuildMux) loadSpec(ctx context.Context, client gwclient.Client) (*dalec
 	m.spec = spec
 
 	return spec, nil
-}
-
-func maybeSetDalecTargetKey(client gwclient.Client, key string) gwclient.Client {
-	opts := client.BuildOpts()
-	if opts.Opts[keyTopLevelTarget] != "" {
-		// do nothing since this is already set
-		return client
-	}
-
-	// optimization to help prevent unnecessary grpc requests
-	// The gateway client will make a grpc request to get the build opts from the gateway.
-	// This just caches those opts locally.
-	// If the client is already a clientWithCustomOpts, then the opts are already cached.
-	if _, ok := client.(*clientWithCustomOpts); !ok {
-		// this forces the client to use our cached opts from above
-		client = &clientWithCustomOpts{opts: opts, Client: client}
-	}
-	return setClientOptOption(client, map[string]string{keyTopLevelTarget: key, "build-arg:" + dalec.KeyDalecTarget: key})
 }
 
 // list outputs the list of targets that are supported by the mux
@@ -312,19 +217,6 @@ func (m *BuildMux) list(ctx context.Context, client gwclient.Client, target stri
 	}
 
 	return ls.ToResult()
-}
-
-type noSuchHandlerError struct {
-	Target    string
-	Available []string
-}
-
-func handlerNotFound(target string, available []string) error {
-	return &noSuchHandlerError{Target: target, Available: available}
-}
-
-func (err *noSuchHandlerError) Error() string {
-	return fmt.Sprintf("no such handler for target %q: available targets: %s", err.Target, strings.Join(err.Available, ", "))
 }
 
 func (m *BuildMux) lookupTarget(ctx context.Context, target string) (matchedPattern string, _ *handler, _ error) {
@@ -507,22 +399,6 @@ func marshalResult[T any](res *gwclient.Result, v *T) error {
 	return nil
 }
 
-// CurrentFrontend is an interface typically implemented by a [gwclient.Client]
-// This is used to get the rootfs of the current frontend.
-type CurrentFrontend interface {
-	CurrentFrontend() (*llb.State, error)
-}
-
-var (
-	_ gwclient.Client = (*clientWithCustomOpts)(nil)
-	_ CurrentFrontend = (*clientWithCustomOpts)(nil)
-)
-
-type clientWithCustomOpts struct {
-	opts gwclient.BuildOpts
-	gwclient.Client
-}
-
 func trimTargetOpt(client gwclient.Client, prefix string) *clientWithCustomOpts {
 	opts := client.BuildOpts()
 
@@ -535,25 +411,6 @@ func trimTargetOpt(client gwclient.Client, prefix string) *clientWithCustomOpts 
 		Client: client,
 		opts:   opts,
 	}
-}
-
-func setClientOptOption(client gwclient.Client, extraOpts map[string]string) *clientWithCustomOpts {
-	opts := client.BuildOpts()
-
-	for key, value := range extraOpts {
-		opts.Opts[key] = value
-	}
-	return &clientWithCustomOpts{
-		Client: client,
-		opts:   opts,
-	}
-}
-
-func (d *clientWithCustomOpts) BuildOpts() gwclient.BuildOpts {
-	return d.opts
-}
-func (d *clientWithCustomOpts) CurrentFrontend() (*llb.State, error) {
-	return d.Client.(CurrentFrontend).CurrentFrontend()
 }
 
 // Handler returns a [gwclient.BuildFunc] that uses the mux to route requests to appropriate handlers
